refactor(database): 重命名日语课程文件并更新app.json背景色
将words38.js、words44.js、words45.js重命名为lesson38.js、lesson44.js、lesson45.js
在app.json中添加backgroundColor配置
