+++ database/everyones_japanese/lesson45.js
